feat(启动检查): 增加Rpc Transaction 默认构造函数检查

+++ QiunetUtils/src/main/java/org/qiunet/utils/scanner/ClassScanner.java
@@ -49,7 +49,7 @@ public final class ClassScanner implements IApplicationContext {
 	/**
 	 * 可以扫描的类型
 	 */
-	private final int scannerTypes;
+	private final long scannerTypes;
 	/**
 	 * 已经回收. 不能再取数据.
 	 */
@@ -57,7 +57,7 @@ public final class ClassScanner implements IApplicationContext {
 
 	private ClassScanner(ScannerType... scannerTypes) {
 		this.reflections = new Reflections("org.qiunet", scanners);
-		int scannerType = 0;
+		long scannerType = 0;
 		for (ScannerType type : scannerTypes) {
 			scannerType |= type.getStatus();
 		}
